fix: TypeScriptからJavaScriptに変更しVercelクラッシュを根本解決
Problem:
- FUNCTION_INVOCATION_FAILED が継続発生
- TypeScript + @vercel/node の依存関係でランタイムエラー
- CommonJS/ES6モジュールの設定不整合

Solution:
- Pure JavaScript (ES6 Module) 実装に変更
- 外部依存なしのシンプルな実装
- Vercel Edge Runtime対応の最小構成
- 基本設計書をHTTP API仕様に更新

Changes:
- api/chat.ts → api/chat.js (TypeScript削除)
- vercel.json でchat.js指定に変更
- 基本設計書でHTTP API仕様を明記
- Stateless実装でサーバーレス環境適応
