+++ docs/02_基本設計.md
@@ -54,7 +54,18 @@ public/
 
 ## 外部インタフェース
 
-### WebSocket API
+### HTTP API（Vercel Serverless Functions）
+
+**重要**: WebSocket APIからHTTP APIベースの実装に変更。
+サーバーレス環境での制約により、HTTP ポーリング方式を採用。
+
+#### API仕様
+- **ベースURL**: `/api/chat`
+- **実装**: JavaScript (ES6 Module)
+- **ランタイム**: Vercel Edge Runtime
+- **状態管理**: Stateless（各リクエストが独立）
+
+### WebSocket API（参考・非実装）
 
 #### 接続イベント
 ```typescript
